# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Novel Dialogue Reader API",
    "description": "轻小说对话辅助阅读器本地后端。契约摘要见 docs/CONTRACTS.md，完整规格见 DEVELOPMENT.md。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/model-profiles": {
      "get": {
        "tags": [
          "model-profiles"
        ],
        "summary": "模型配置列表（不含密钥）",
        "operationId": "list_profiles_route_api_model_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_list_ModelProfileOut__"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "model-profiles"
        ],
        "summary": "新建模型配置",
        "operationId": "create_profile_route_api_model_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ModelProfileOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/model-profiles/{profile_id}": {
      "patch": {
        "tags": [
          "model-profiles"
        ],
        "summary": "更新配置（含密钥 keep/replace/remove）",
        "operationId": "update_profile_route_api_model_profiles__profile_id__patch",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProfilePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ModelProfileOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "model-profiles"
        ],
        "summary": "删除配置与凭据引用",
        "operationId": "delete_profile_route_api_model_profiles__profile_id__delete",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/model-profiles/test": {
      "post": {
        "tags": [
          "model-profiles"
        ],
        "summary": "连接测试（微型结构化请求；不评估小说效果）",
        "description": "有预算的微型请求：检查鉴权与输出可解析。\n\n网络调用**不在数据库事务里**：先写入 PREPARED 的推理尝试记录，再调用提供方，\n最后回写结果；usage 缺失时保持 unknown，不写成 0。",
        "operationId": "test_connection_route_api_model_profiles_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionTestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ConnectionTestOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionTestIn": {
        "properties": {
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelProfileCreate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ConnectionTestIn",
        "description": "连接测试请求：已保存配置，或临时草稿（草稿不落库）。"
      },
      "ConnectionTestOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "adapter": {
            "type": "string",
            "title": "Adapter"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage"
          },
          "usage_unknown": {
            "type": "boolean",
            "title": "Usage Unknown",
            "default": true
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ok",
          "protocol",
          "model",
          "adapter",
          "detail",
          "run_id"
        ],
        "title": "ConnectionTestOut"
      },
...
      "CredentialMode": {
        "type": "string",
        "enum": [
          "session",
          "system",
          "none"
        ],
        "title": "CredentialMode"
      },
...
      "DataEnvelope_ConnectionTestOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConnectionTestOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[ConnectionTestOut]"
      },
...
      "DataEnvelope_ModelProfileOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ModelProfileOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[ModelProfileOut]"
      },
...
      "DataEnvelope_list_ModelProfileOut__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ModelProfileOut"
            },
            "type": "array",
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[list[ModelProfileOut]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelProfileCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "chat-completions-compatible"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "description": "API 根路径，例如 https://api.example.com/v1"
          },
          "model": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Model"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "description": "生成参数（不能含密钥字段）"
          },
          "credential_mode": {
            "$ref": "#/components/schemas/CredentialMode",
            "default": "session"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "只在请求里出现，绝不回传"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "base_url",
          "model"
        ],
        "title": "ModelProfileCreate"
      },
      "ModelProfileOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "credential_mode": {
            "$ref": "#/components/schemas/CredentialMode"
          },
          "has_key": {
            "type": "boolean",
            "title": "Has Key"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "credential_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Warning",
            "description": "例如系统凭据库不可用、已降级为会话密钥"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "protocol",
          "base_url",
          "model",
          "credential_mode",
          "has_key",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "ModelProfileOut"
      },
      "ModelProfilePatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "credential_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CredentialMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "提供则视为替换密钥"
          },
          "remove_api_key": {
            "type": "boolean",
            "title": "Remove Api Key",
            "description": "清除该配置的密钥",
            "default": false
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ModelProfilePatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}